pulled emotions sass out of shared and into its own partial, simplified it.
removed javascript to show / hide the emotions panel, now handled entirely through sass on hover. this resolved the z-indexing on the dream stream. and less JS is always better.

re-worked the styling for the 1d thumbs, now much more dynamic and any amount of content can be thrown in, expanded, etc.